Fix integer overflow crashing paging on a far-past-the-end page
size was already capped at 200, but page was only bounded by
Integer.MAX_VALUE, and the offset was computed as (page - 1) * size in int.
?page=11000001&size=200 overflowed to -2094967296, reached subList as a
negative fromIndex and returned a 500.

Computing the offset in long fixes it at the source; both bounds are then
clamped to the match count, so narrowing back to int is safe. A page past
the end stays what it was for smaller values -- 200 with an empty data array,
not an error -- since asking for a page that doesn't exist yet is a legitimate
request, not a malformed one. No artificial cap on page was added: the crash
was the arithmetic, not the input.

+++ src/main/java/com/ata/jobdata/service/JobDataService.java
@@ -32,8 +32,11 @@ public class JobDataService {
             matched = matched.stream().sorted(sort).toList();
         }
 
-        int from = Math.min((params.page() - 1) * params.size(), matched.size());
-        int to = Math.min(from + params.size(), matched.size());
+        // In long: page * size exceeds int for a far-past-the-end page, and wrapping negative used to
+        // reach subList as a negative fromIndex. Both bounds are clamped to the match count, so the
+        // narrowing back to int is safe.
+        int from = (int) Math.min((long) (params.page() - 1) * params.size(), matched.size());
+        int to = (int) Math.min((long) from + params.size(), matched.size());
         List<Map<String, Object>> data = matched.subList(from, to).stream()
                 .map(record -> project(record, params.fields()))
                 .toList();
